Autosync:  title     = {Synopsis of the Go Programming Language},
- year          = {2018},
+ year          = {2023},

- eprint        = {arXiv:1500034},
+ eprint        = {arXiv:2306.09730},

- archiveprefix = {arXiv},
+ archiveprefix = {arXiv},

- primaryclass  = {cs.DC},
+ primaryclass  = {cs.CL},

- url           = {http://golang.org/doc/}
+ url           = {https://arxiv.org/abs/2306.09730}
}

The new BibTeX entry you provided updates the year of the paper to 2023 and changes the title of the paper to "Synopsis of the Go Programming Language". The rest of the fields remain the same as the original entry, with the URL being updated to point to the arXiv abstract page for the new version of the paper.